add lambda authorization

+++ lambda/src/main/java/lambda/test/DeleteTest.java
@@ -2,17 +2,23 @@ package lambda.test;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import lambda.Handler;
+import model.request.AuthenticatedRequest;
 import model.request.TestRequest;
 import model.test.Test;
 import util.Response;
 
-public class DeleteTest extends Handler<TestRequest> {
+public class DeleteTest extends Handler<AuthenticatedRequest<TestRequest>> {
 
     @Override
-    public Response handleRequest(TestRequest input, Context context) {
+    public Response handleRequest(AuthenticatedRequest<TestRequest> authenticatedRequest, Context context) {
+        if(!authenticatedRequest.isRecruiter())
+            return new Response(403, "Recruiter permissions required");
+        if(!authenticatedRequest.getUserId().equals(authenticatedRequest.getBody().getOwnerId()))
+            return new Response(403, "Insufficient permissions");
+
         Test test = new Test();
-        test.setRecruiterId(input.getOwnerId());
-        test.setTestId(input.getTestId());
+        test.setRecruiterId(authenticatedRequest.getBody().getOwnerId());
+        test.setTestId(authenticatedRequest.getBody().getTestId());
         getMapper().delete(test);
         return new Response(204, "");
     }
